Final Project/requirements.txt
python==3.9.5
numpy==1.16.6
pytorch==1.11.0
torchvision==0.12.0
matplotlib==3.3.4
opencv==4.5.4
numpy==1.16.6
tqdm==4.61.1
